requirements.txt
# Production dependencies
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pydantic[email]>=1.10.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
PyJWT>=2.0.0